services/api-fastapi/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pydantic==2.8.2
pyyaml==6.0.2
python-multipart==0.0.9
